src/catalog/pyproject.toml
[project]
name = "catalog"
version = "0.1.0"
description = "blah"
authors = [{ name = "Mike Waters", email = "mike@mikewaters.net" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "marimo==0.15.5",
    "pydantic-settings>=2.11.0",
    "loguru>=0.7.0",
    "llama-index-core>=0.14.13",
    "langfuse>=3.12.1",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "llama-index-vector-stores-qdrant>=0.9.1",
    "qdrant-client>=1.16.2",
    "typer>=0.21.1",
    "agentlayer",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = [
  "catalog",
  "index",
]

[tool.pytest.ini_options]
addopts = [
    "--ignore=catalog/tests/idx/scenario"
]

[project.scripts]
catalog = "catalog.cli:main"

[dependency-groups]
dev = [
    "agentlayer>=0.1",
    "pytest>=9.0.2",
    "pytest-testmon>=2.2.0",
]
